Rename some vars; redefine voxel datatype

+++ src/granular/GranularDefines.h
@@ -5,9 +5,11 @@ namespace sgps {
 #define N_MANUFACTURED_ITEMS 3
 #define N_INPUT_ITEMS 4
 
-typedef long int voxelID_ts;
+typedef unsigned int voxelID_ts;
 typedef unsigned int bodyID_t;
 typedef unsigned char radiusIndex_t;
 typedef unsigned char materialIndex_t;
+
+// somehow add array materialsArray and radiiArray??
 
 } // namespace sgps
